% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_events.R
\name{align_events}
\alias{align_events}
\title{Aligns data by events}
\usage{
align_events(df, df_isevent, dovars, leng_threshold, before, after, nbins)
}
\arguments{
\item{df}{A data frame containing all data continuously along time, required column named \code{site, date}.}

\item{df_isevent}{A data frame \code{nrow(df_isevent)==nrow(df)} specifying whether respective dates
(matching dates in \code{df}), fall satisfy a condition that is used to define events. Events are
consecutive dates, where this condition is satisfied (minimum length for defining an event is given by
\code{leng_threshold}). Required columns \code{site, date}}

\item{dovars}{A vector of character strings specifying which columns of \code{df} to re-arrange.}

\item{before}{An integer specifying the number of days before the event onset to be retained in re-arranged data}

\item{after}{An integer specifying the number of days after the event onset to be retained in re-arranged data}
}
\value{
An aligned data frame
}
\description{
Uses a vector specifying whether data falls into an event to reshape data, aligning by the onset of the event
}
\examples{
df_alg <- align_events( df, truefalse, before=30, after=300 )

}
