% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VI.R
\name{calc_VI}
\alias{calc_VI}
\title{Batch calculate VI}
\usage{
calc_VI(df, indices = "data/spectral-indices-table.csv")
}
\arguments{
\item{df}{the machine learning data frame including the 7 MODIS NBAR bands}

\item{indices}{path to the spectral indices file}
}
\value{
select spectral indices for all values in the data frame
}
\description{
Calculate VI for the machine learning input
data frame based on the spectral indices table
(in csv format).
}
