% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_flue.R
\name{index_flue}
\alias{index_flue}
\title{Creates fLUE drought index values}
\usage{
index_flue(df)
}
\arguments{
\item{df}{a machine learning data frame}
}
\value{
the same data frame with numbered
drought instances per site, as well as the consecutive
days the drought lasted, as well as the preceding
20 days if no overlap with other droughts is present
}
\description{
Indexes periods of drought consecutively
per site (idx), and provides an a numeric
value in days that the drought lasted
preceding days are also counted (if not
overlapping with a previous drought for
comparison purposes)
}
