% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_normalized_differences.R
\name{merge_normalized_differences}
\alias{merge_normalized_differences}
\title{Merge / calculate difference ratios}
\usage{
merge_normalized_differences(ml_df)
}
\arguments{
\item{ml_df}{a machine learning data frame with "sur_refl" prefix column
names apply the band ratios to}
}
\value{
A machine learning data frame with normalized band ratios appended
}
\description{
Calculate normalized difference ratios
for all possible "sur_refl" band combinations
}
