% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{normalized_difference}
\alias{normalized_difference}
\title{Normalized difference}
\usage{
normalized_difference(band_1, band_2)
}
\arguments{
\item{band_1}{first band}

\item{band_2}{second band}
}
\value{
normalized difference index for two bands
}
\description{
Calculates normalized difference for two (arbitrary) bands. The difference
is taken as band 1 - band 2
}
