% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ml_data.R
\name{read_ml_data}
\alias{read_ml_data}
\title{Read in machine learning dataframe}
\usage{
read_ml_data(path, spatial = FALSE)
}
\arguments{
\item{path}{path to the raw machine learing RDS file}

\item{spatial}{TRUE or FALSE, restrict output for spatial scaling}
}
\value{
data frame with machine learning features and targets
}
\description{
subsets data for spatial scaling or not
}
